[vom] Create and restore backup
- create backup of running datastore
- restore running datastore affter sweetcomb crash

Notes:
This solution is not ideal. It only show how the backup can work and how can
we use it.
Ideal solution will be if we don't use system function, but use
directory function from sysrepo. But this implementation should take few
weeks and we probably need also path sysrepo.

Change-Id: Ifb483afca07fabb2b8fd7e5f18008a262190228e
Signed-off-by: Andrej Kozemcak <andrej.kozemcak@pantheon.tech>